sending geo data to front end, displaying it as markers, and taken out schema defs from index into models folder
