+++ web/public/javascripts/map.js
@@ -12,10 +12,11 @@ L.tileLayer('https://stamen-tiles-{s}.a.ssl.fastly.net/toner-lite/{z}/{x}/{y}{r}
 	ext: 'png'
 }).addTo(map);
 
-// Create point feature for somewhere downtown
-var myDataPoint = L.marker([43.6529, -79.3849]).addTo(map);
-
-// Bind popup to Data Point object
-myDataPoint.bindPopup("<h3>City Hall (not really)</h3><p>Toronto, ON<br>Information about city hall.</p>");
-
-map.setZoom(13);
+L.geoJson(dumbpoints, {
+	pointToLayer: function (feature, latlng) {
+		//return L.circleMarker(latlng);
+		return L.marker(latlng);
+	}
+}).bindPopup(function (layer) {
+	return layer.feature.geometry.another;
+}).addTo(map)
