+++ web/views/map.pug
@@ -1,11 +1,11 @@
 extends layout.pug
 block content
-    #sidebar
-        h1 This is a sidebar
-    #map
+	#sidebar
+		h1 This is a sidebar
+	#map
 
-    script(type='text/javascript').
-        var map = L.map('map').setView([#{lat},#{lng}], 15);
-    script(src='/javascripts/map.js')
-
-
+	script(type='text/javascript').
+		var map = L.map('map').setView([#{lat},#{lng}], 13);
+		var dumbpoints = !{JSON.stringify(points)};
+		//console.log(dumbpoints);
+	script(src='/javascripts/map.js')
